decouple receiver and state processing
The issue with this design is that server has to send both the state information as well as the machine.
Also for each state info received from the server we need a queue as a middle man to fetch and put data. We can avoid this
extra storage by running listeners for each subscription